Moved the SQLInputField component out of the services folder and into the component folder.
