+++ packages/front-end/components/Settings/EditDataSource/ExperimentAssignmentQueries/AddEditExperimentAssignmentQueryModal.tsx
@@ -9,7 +9,7 @@ import cloneDeep from "lodash/cloneDeep";
 import uniqId from "uniqid";
 import Field from "../../../Forms/Field";
 import StringArrayField from "../../../Forms/StringArrayField";
-import SQLInputField from "../../../../services/SQLInputField";
+import SQLInputField from "../../../SQLInputField";
 
 type EditExperimentAssignmentQueryProps = {
   exposureQuery?: ExposureQuery;
